Removed version 10 when installing Node.js
This was added in for testing, as the current Appveyor docs say that "lts" is version 8 and not 10. However, version 10 does seem to be installed when "lts" is supplied.

This also helps provide future version compatibility when new versions of Node.js are released without needing to update the `.appveyor.yml` configuration file.

The Node.js order has also been reversed, so that it is inline with that of `.travis.yml` when viewing CI build output files.